Wave 37: 3260 -> 3431 (+171), bytes 41.95% -> 47.35%
171 functions verified by exit code out of 203 attempted, across 25 agents.
Both builds reproduce SHA1 14dd0b22c894865867aff89e8116b2dffae25605:
`make SPLIT=1 compare` and `make compare`, run and read separately.

THE LOOP BAND IS NOT EXPENSIVE, AND THE SCREEN HAS BEEN HIDING MOST OF THE
CORPUS SINCE WAVE 16. overlap_screen.py filters `backward_branches == 0` on
the theory that a loop is one COUPLED decision where a wrong shape corrupts
every later byte. Seven agents across six blocks measured 39 of 41 (95%) at
62-166 bytes/attempt -- at or above the straight-line rate. Every miss pointed
at itself; loop-body errors stayed confined while prologue, guard and tail
stayed byte-exact. 869 clean loop functions / 227,912 bytes sit behind that
filter against 244 / 41,078 in front of it. Wave 38 should batch there first.

Loops need no special authoring but do need read-out rules straight-line code
never exercises; five are new in docs/agbcc-codegen.md, the most useful being
that the preheader is written by three passes in a fixed order (source, LICM
hoists, then strength_reduce's biv/giv inits), so anything after a hoisted
invariant was written by the loop optimiser and must not be authored.

Batch size capped at 5-6 functions per agent (standing user directive, now in
SKILL.md). Every batch under the cap finished; the three that lost functions to
context exhaustion were 12-15 wide and left five stub work/ dirs behind.

Six rules were REFUTED or narrowed by controlled probe rather than dropped:
  - "a low-scoring draft is safe to inherit"      9% vs 83% for undrafted
  - "`?:` on a store is not real"                 a promoted file does it
  - "ldrh;lsls;asrs means an (s16) cast"          sign_extend always folds to ldrsh
  - "bgt/bge on a decrementing counter"           check_dbra_loop leaves that too
  - "reuse, not a fresh name"                     inverts; decide by the exit test
  - "the permuter gets nothing on allocation"     order- and slot-wrong are its cases
Each is annotated in place; earlier replications are flagged, not deleted.

Fixed three things that would each have broken the build:
  - two functions regressed by a later agent widening a struct member they read
    from work/ (caught only by the final from-scratch sweep -- 5th such break)
  - struct MapScreen defined 17 different ways across drafts; unified, all 14
    promotable readers re-verified byte-for-byte
  - gUnknown_084A077C referenced but carved out of no data block -- SPLIT-only,
    the assembly build prints OK beside that failure

+++ src/decomp/c_080129F8.c
@@ -15,7 +15,10 @@
  * `movs #0; b` / `movs #1` split across the pool is the if/else-return form,
  * not a returned comparison -- a returned comparison would preset the false
  * value and arrive via `adds r0, r1, #0` with no unconditional branch. */
-int sub_080129F8(u16 a)
+/* Wave 37 (W37-H): retyped `int` -> `u8`. Its only caller sub_08035170 narrows
+ * the result with `lsls r0,#0x18; cmp r0,#0`, which an int return does not
+ * produce. Byte-exact before and after -- both arms return a 0/1 constant. */
+u8 sub_080129F8(u16 a)
 {
     if (sub_080129E0() % 10000 < a * 100)
         return 1;
